flask-revolut-donate/requirements.txt
Flask
Flask-Cors
Flask-Mail
requests
python-dotenv